+++ CHANGELOG.md
@@ -6,7 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
-## [Unreleased]
+## [v0.4.2] - 2020-06-21
 
 ### Fixed
 
@@ -15,6 +15,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Definition of private tagged tuples.
+- `error/0` definitions for `Tagged.Status`
+
+### Changed
+
+- The `|> with_...()` functions are now actually functions, instead of macros.
 
 ## [v0.4.1] - 2020-05-31
 
@@ -80,7 +85,8 @@ Initial package release.
 - Generate a constructor / destructuring macro for a tag.
 - Generate a typedef for a tag.
 
-[Unreleased]: https://github.com/notCalle/elixir-tagged/compare/v0.4.1..HEAD
+[Unreleased]: https://github.com/notCalle/elixir-tagged/compare/v0.4.2..HEAD
+[v0.4.2]: https://github.com/notCalle/elixir-tagged/releases/tag/v0.4.2
 [v0.4.1]: https://github.com/notCalle/elixir-tagged/releases/tag/v0.4.1
 [v0.4.0]: https://github.com/notCalle/elixir-tagged/releases/tag/v0.4.0
 [v0.3.0]: https://github.com/notCalle/elixir-tagged/releases/tag/v0.3.0